Development frontend scales (#78)
* Testing mobile repository

Testing Ryan Norman's mobile repository setup.

* Spring environment

Establishing the environment to develop in.

* Cleanup

removing testfile

* random changes

commiting random changes

* Utilizing the new request system

Created a request system. not complete yet, as unit tests do not use post correctly.

* Auto stash before merge of "development-database-garrison" and "origin/development-database-garrison"

* All files for Registration and User

* Removing bad file

* resolving small indent error

* resolving small indent error

* Rebuilding the unit test

Not passing yet

* adding travis CI

adding travis CI to my development build

* rename travis.yml to .travis.yml

* updating travis.yml for frontend testing

* Login and Usernames 

Login and user names with Maven and SQL compat - AS well as testable jsp

* added registration

added registration modal from the login page

* added stashed changes

* applying stashed changes

* Update login.component.ts

* added karma test runner with basic unittests per component

* testing karma and travis ci

* testing karma and travis ci

* testing karma and travis ci

* testing karma and travis ci

* testing karma and travis ci

* testing completed yml

* Comment class created

The comment class will be the framework built for text comments.

* Update .travis.yml

* testing completed yml

* Created the Comment History Class

The framework for displaying and managing the user's history. This is only for comments, not posts. This will need to retrieve its data from the database.

* Created a reset password function

In user, there is now a function to set the password to a random 8- digit number.

* Framework for image histories

Created a framework for a user's image history.

* updated dependencies for sidebar

* Created a visibility bool for users

Created a boolean variable for public/private accounts.

* working login

* final development integration